+++ src/TinyIR.h
@@ -34,9 +34,9 @@
  * @{
  */
 
-#define VERSION_TINYIR "2.1.0"
+#define VERSION_TINYIR "2.2.0"
 #define VERSION_TINYIR_MAJOR 2
-#define VERSION_TINYIR_MINOR 1
+#define VERSION_TINYIR_MINOR 2
 #define VERSION_TINYIR_PATCH 0
 // The change log is at the bottom of the file
 
@@ -243,7 +243,7 @@ struct TinyIRReceiverCallbackDataStruct {
     uint8_t Command;
 #endif
     uint8_t Flags; // Bit coded flags. Can contain one of the bits: IRDATA_FLAGS_IS_REPEAT and IRDATA_FLAGS_PARITY_FAILED
-    bool justWritten; ///< Is set true if new data is available. Used by the main loop, to avoid multiple evaluations of the same IR frame.
+    bool justWritten; ///< Is set true if new data is available. Used by the main loop / TinyReceiverDecode(), to avoid multiple evaluations of the same IR frame.
 };
 extern volatile TinyIRReceiverCallbackDataStruct TinyIRReceiverData;
 
@@ -252,6 +252,7 @@ bool initPCIInterruptForTinyReceiver();
 bool enablePCIInterruptForTinyReceiver();
 void disablePCIInterruptForTinyReceiver();
 bool isTinyReceiverIdle();
+bool TinyReceiverDecode();
 void printTinyReceiverResultMinimal(Print *aSerial);
 
 void sendFAST(uint8_t aSendPin, uint16_t aCommand, uint_fast8_t aNumberOfRepeats = 0);
@@ -263,6 +264,9 @@ void sendNEC(uint8_t aSendPin, uint16_t aAddress, uint16_t aCommand, uint_fast8_
 void sendExtendedNEC(uint8_t aSendPin, uint16_t aAddress, uint16_t aCommand, uint_fast8_t aNumberOfRepeats = 0, bool aSendNEC2Repeats = false);
 
 /*
+ *  Version 2.2.0 - 7/2024
+ *  - New TinyReceiverDecode() function to be used as drop in for IrReceiver.decode().
+ *
  *  Version 2.1.0 - 2/2024
  *  - New sendExtendedNEC() function and new parameter aSendNEC2Repeats.
  *