+++ changelog.md
@@ -3,6 +3,9 @@ The latest version may not be released!
 See also the commit log at github: https://github.com/Arduino-IRremote/Arduino-IRremote/commits/master
 # 4.4.1
 - Support for ESP 3.0 by akellai.
+- restartTimer() now uses variable sMicrosAtLastStopTimer to keep track of uncounted ticks between stopTimer() and restartTimer().
+- Removed functions addTicksToInternalTickCounter() and addMicrosToInternalTickCounter(), which were added in 4.1.0.
+- Version 2.2.0 of TinyIR with new TinyReceiverDecode() function to be used as drop in for IrReceiver.decode().
 
 # 4.4.0
 - Using 8 bit raw timing buffer for all timings except frame gap (former rawbuf[0]).